requirements.txt
Flask
gunicorn
requests
prometheus_client
python-dotenv
pandas
aiohttp
